+++ app/models/emailbutler/message.rb
@@ -2,5 +2,12 @@
 
 module Emailbutler
   class Message < ApplicationRecord
+    after_initialize :generate_uuid
+
+    private
+
+    def generate_uuid
+      self.uuid ||= SecureRandom.uuid
+    end
   end
 end