+++ config/routes.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 Emailbutler::Engine.routes.draw do
+  post '/webhooks', to: 'webhooks#create'
 end