+++ db/migrate/20220901203215_add_uuid_to_messages.rb
@@ -1,0 +1,7 @@
+class AddUuidToMessages < ActiveRecord::Migration[7.0]
+  def change
+    enable_extension 'pgcrypto'
+    enable_extension 'uuid-ossp'
+    add_column :emailbutler_messages, :uuid, :uuid, null: false, index: true
+  end
+end